+++ service/src/main/java/org/ehrbase/dao/access/jooq/EntryAccess.java
@@ -26,6 +26,7 @@ import com.nedap.archie.rm.archetyped.TemplateId;
 import com.nedap.archie.rm.composition.*;
 import com.nedap.archie.rm.datatypes.CodePhrase;
 import com.nedap.archie.rm.datavalues.DvCodedText;
+import com.nedap.archie.rm.datavalues.DvText;
 import com.nedap.archie.rm.generic.PartyProxy;
 import com.nedap.archie.rm.support.identification.ArchetypeID;
 import com.nedap.archie.rm.support.identification.ObjectVersionId;
@@ -43,8 +44,11 @@ import org.ehrbase.ehr.knowledge.I_KnowledgeCache;
 import org.ehrbase.jooq.pg.enums.EntryType;
 import org.ehrbase.jooq.pg.tables.records.EntryHistoryRecord;
 import org.ehrbase.jooq.pg.tables.records.EntryRecord;
+import org.ehrbase.jooq.pg.udt.records.DvCodedTextRecord;
 import org.ehrbase.serialisation.dbencoding.RawJson;
 import org.ehrbase.service.IntrospectService;
+import org.ehrbase.service.RecordedDvCodedText;
+import org.ehrbase.service.RecordedDvText;
 import org.jooq.*;
 import org.jooq.impl.DSL;
 
@@ -64,7 +68,7 @@ public class EntryAccess extends DataAccess implements I_EntryAccess {
     public static final String DB_INCONSISTENCY = "DB inconsistency:";
 
     private EntryRecord entryRecord;
-    private I_ContainmentAccess containmentAccess;
+//    private I_ContainmentAccess containmentAccess;
 
     private Composition composition;
 
@@ -249,6 +253,9 @@ public class EntryAccess extends DataAccess implements I_EntryAccess {
                 case CONTEXT:
                     composition.setContext((EventContext) systemValue.getValue());
                     break;
+                case NAME:
+                    composition.setName((DvText) systemValue.getValue());
+                    break;
 
                 default:
                     throw new IllegalArgumentException("Could not handle composition attribute:" + systemValue.getKey());
@@ -292,7 +299,6 @@ public class EntryAccess extends DataAccess implements I_EntryAccess {
 
         RawJson rawJson = new RawJson();
         record.setEntry(JSONB.valueOf(rawJson.marshal(composition)));
-        containmentAccess = new ContainmentAccess(getDataAccess(), record.getId(), record.getArchetypeId(), rawJson.getLtreeMap(), true);
     }
 
     /**
@@ -315,6 +321,8 @@ public class EntryAccess extends DataAccess implements I_EntryAccess {
 
         setCompositionFields(entryRecord, composition);
 
+        setCompositionName(composition.getName());
+
         this.composition = composition;
     }
 
@@ -326,31 +334,18 @@ public class EntryAccess extends DataAccess implements I_EntryAccess {
     @Override
     public UUID commit(Timestamp transactionTime) {
 
-        //--------------- TODO: WIP refactoring it in jooq
-        /*InsertQuery<?> insertQuery = context.insertQuery(ENTRY);
-        insertQuery.addValue(ENTRY.SEQUENCE, entryRecord.getSequence());
-        insertQuery.addValue(ENTRY.COMPOSITION_ID, entryRecord.getCompositionId());
-        insertQuery.addValue(ENTRY.TEMPLATE_ID, entryRecord.getTemplateId());
-        insertQuery.addValue(ENTRY.ITEM_TYPE, entryRecord.getItemType());
-        insertQuery.addValue(ENTRY.ARCHETYPE_ID, entryRecord.getArchetypeId());
-        insertQuery.addValue(ENTRY.CATEGORY, entryRecord.getCategory());
-        insertQuery.addValue(ENTRY.ENTRY_, DSL.val(getEntryJson() + "::jsonb"));
-        insertQuery.addValue(ENTRY.SYS_TRANSACTION, transactionTime);
-
-        int result;
-        try {
-            result = insertQuery.execute();
-        } catch (DataAccessException e) {
-            throw new InternalServerException("Problem executing database operation", e);
-        }
-
-        if (result < 1) // TODO check result for successful execution -> is '< 1' correct as condition?
-            throw new InternalServerException("Entry commit failed");*/
-        //------------- END --------------------
-
-        //use jOOQ!
+        //use jOOQ
         Record result = getContext()
-                .insertInto(ENTRY, ENTRY.SEQUENCE, ENTRY.COMPOSITION_ID, ENTRY.TEMPLATE_ID, ENTRY.ITEM_TYPE, ENTRY.ARCHETYPE_ID, ENTRY.CATEGORY, ENTRY.ENTRY_, ENTRY.SYS_TRANSACTION)
+                .insertInto(ENTRY,
+                        ENTRY.SEQUENCE,
+                        ENTRY.COMPOSITION_ID,
+                        ENTRY.TEMPLATE_ID,
+                        ENTRY.ITEM_TYPE,
+                        ENTRY.ARCHETYPE_ID,
+                        ENTRY.CATEGORY,
+                        ENTRY.ENTRY_,
+                        ENTRY.SYS_TRANSACTION,
+                        ENTRY.NAME)
                 .values(DSL.val(getSequence()),
                         DSL.val(getCompositionId()),
                         DSL.val(getTemplateId()),
@@ -358,14 +353,15 @@ public class EntryAccess extends DataAccess implements I_EntryAccess {
                         DSL.val(getArchetypeId()),
                         DSL.val(getCategory()),
                         DSL.val(getEntryJson()),
-                        DSL.val(transactionTime))
+                        DSL.val(transactionTime),
+                        DSL.val(getCompositionName()))
                 .returning(ENTRY.ID)
                 .fetchOne();
 
-        if (containmentAccess != null) {
-            containmentAccess.setCompositionId(entryRecord.getCompositionId());
-            containmentAccess.update();
-        }
+//        if (containmentAccess != null) {
+//            containmentAccess.setCompositionId(entryRecord.getCompositionId());
+//            containmentAccess.update();
+//        }
 
         return result.getValue(ENTRY.ID);
         //return entryRecord.getId(); // TODO: part of WIP refactoring from above
@@ -408,15 +404,16 @@ public class EntryAccess extends DataAccess implements I_EntryAccess {
         updateQuery.addValue(ENTRY.CATEGORY, DSL.field(DSL.val(getCategory())));
         updateQuery.addValue(ENTRY.ENTRY_, DSL.field(DSL.val(getEntryJson())));
         updateQuery.addValue(ENTRY.SYS_TRANSACTION, DSL.field(DSL.val(transactionTime)));
+        updateQuery.addValue(ENTRY.NAME, DSL.field(DSL.val(getCompositionName())));
         updateQuery.addConditions(ENTRY.ID.eq(getId()));
 
 
         log.debug("Update done...");
 
-        if (containmentAccess != null) {
-            containmentAccess.setCompositionId(entryRecord.getCompositionId());
-            containmentAccess.update();
-        }
+//        if (containmentAccess != null) {
+//            containmentAccess.setCompositionId(entryRecord.getCompositionId());
+//            containmentAccess.update();
+//        }
 
         return updateQuery.execute() > 0;
     }
@@ -463,6 +460,14 @@ public class EntryAccess extends DataAccess implements I_EntryAccess {
     @Override
     public UUID getCategory() {
         return entryRecord.getCategory();
+    }
+
+    public DvCodedTextRecord getCompositionName() {
+        return entryRecord.getName();
+    }
+
+    public void setCompositionName(DvText compositionName){
+        new RecordedDvText().toDB(entryRecord, ENTRY.NAME, compositionName);
     }
 
     @Override